example_server/pyproject.toml
# This project is configured for UV package manager
# Install with: uv sync
# Run with: uv run <command>

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "example_server"
version = "0.1.0"
description = "Example MCP server showcasing SAAGA decorator patterns"
authors = [
    {name = "SAAGA Team", email = "team@saaga.ai"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp>=1.0.0",
    "mcp[cli]>=1.0.0",
    "anyio>=4.5",
    "starlette>=0.36.0",
    "uvicorn>=0.27.0",
    "click>=8.0.0",
    "platformdirs>=3.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-multipart>=0.0.6",
    "loguru>=0.7.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",  # Required for test cleanup
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "streamlit>=1.28.0",
    "plotly>=5.0.0",
    "pandas>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
example_server-server = "example_server.server.app:main"
test-mcp-integration = "example_server.tests.integration.cli:run_tests_entry"

[project.urls]
Homepage = "https://github.com/SAAGA Team/example_server"
Documentation = "https://github.com/SAAGA Team/example_server/blob/main/README.md"
Repository = "https://github.com/SAAGA Team/example_server"
Issues = "https://github.com/SAAGA Team/example_server/issues"

[tool.setuptools]
packages = ["example_server", "tests"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=example_server --cov-report=html --cov-report=term-missing"